refactor: decompose god modules (feature_engineering, frankenstein_editor, folder_tool_processing)
- Split feature_engineering.py (1191 lines) into 4 focused modules:
  - feature_types.py: Shared enums and dataclasses
  - feature_extractor.py: FeatureExtractor class
  - feature_selector.py: FeatureSelector class
  - feature_transformer.py: FeatureTransformer class
  - feature_engineering.py: Thin facade for backward compatibility

- Split frankenstein_editor.py (1400 lines) into 4 focused modules:
  - editor_types.py: Shared data types
  - editor_clipboard.py: ClipboardMixin (copy operations)
  - editor_modifications.py: ModificationMixin (delete/rename/attach/mirror)
  - frankenstein_editor.py: Main class composing mixins

- Split folder_tool_processing.py (984 lines) into 4 focused modules:
  - folder_tool_analysis.py: AnalysisMixin (report generation, validation)
  - folder_tool_archive.py: ArchiveMixin (ZIP creation)
  - folder_tool_ui_processing.py: UIProcessingMixin (dialogs, progress)
  - folder_tool_processing.py: Slim facade composing all mixins

All original APIs preserved via facades and re-exports.
