+++ src/shared/python/model_generation/editor/editor_modifications.py
@@ -1,0 +1,603 @@
+"""Modification operations mixin for the Frankenstein Editor.
+
+Provides link/joint deletion, renaming, modification, attachment/detachment,
+batch prefix application, and subtree mirroring.
+"""
+
+from __future__ import annotations
+
+import logging
+from typing import Any
+
+from model_generation.core.types import Joint, JointType, Link, Origin
+
+logger = logging.getLogger(__name__)
+
+
+class ModificationMixin:
+    """Modification operations for the FrankensteinEditor.
+
+    Expects the host class to provide:
+    - self._models: dict[str, ParsedModel]
+    - self._rename_callbacks: list[Callable]
+    - self._clipboard: list[...]
+    - self._save_state() -> None
+    - self.get_connecting_joint(model_id, link_name) -> Joint | None
+    - self.copy_subtree(model_id, root_link) -> bool
+    - self._generate_unique_name(base_name, existing_names) -> str
+    """
+
+    # ============================================================
+    # Direct Modifications
+    # ============================================================
+
+    def delete_link(
+        self,
+        model_id: str,
+        link_name: str,
+        reparent_children: bool = True,
+    ) -> bool:
+        """
+        Delete a link from a model.
+
+        Args:
+            model_id: Target model
+            link_name: Link to delete
+            reparent_children: If True, attach children to grandparent
+
+        Returns:
+            True if deleted
+        """
+        model = self._models.get(model_id)
+        if not model:
+            logger.error(f"Model '{model_id}' not found")
+            return False
+
+        if model.read_only:
+            logger.error(f"Model '{model_id}' is read-only")
+            return False
+
+        link = model.get_link(link_name)
+        if not link:
+            logger.error(f"Link '{link_name}' not found")
+            return False
+
+        self._save_state()
+
+        # Get parent
+        parent_name = model.get_parent(link_name)
+        parent_joint = self.get_connecting_joint(model_id, link_name)
+
+        # Get children
+        children = model.get_children(link_name)
+
+        # Reparent children if requested
+        if reparent_children and parent_name:
+            for child in children:
+                child_joint = self.get_connecting_joint(model_id, child)
+                if child_joint:
+                    child_joint.parent = parent_name
+                    # Adjust origin if we have parent joint info
+                    if parent_joint:
+                        # Combine transforms (simplified - just add positions)
+                        px, py, pz = parent_joint.origin.xyz
+                        cx, cy, cz = child_joint.origin.xyz
+                        child_joint.origin = Origin(
+                            xyz=(px + cx, py + cy, pz + cz),
+                            rpy=child_joint.origin.rpy,
+                        )
+        elif not reparent_children:
+            # Delete children recursively
+            for child in children:
+                self.delete_subtree(model_id, child)
+
+        # Remove the link
+        model.links = [link for link in model.links if link.name != link_name]
+
+        # Remove connecting joint
+        model.joints = [j for j in model.joints if j.child != link_name]
+
+        logger.info(f"Deleted link '{link_name}' from '{model_id}'")
+        return True
+
+    def delete_subtree(self, model_id: str, root_link: str) -> bool:
+        """
+        Delete a subtree (link and all descendants).
+
+        Args:
+            model_id: Target model
+            root_link: Root link of subtree to delete
+
+        Returns:
+            True if deleted
+        """
+        model = self._models.get(model_id)
+        if not model:
+            logger.error(f"Model '{model_id}' not found")
+            return False
+
+        if model.read_only:
+            logger.error(f"Model '{model_id}' is read-only")
+            return False
+
+        subtree = model.get_subtree(root_link)
+        if not subtree:
+            logger.error(f"Link '{root_link}' not found")
+            return False
+
+        self._save_state()
+
+        # Remove all links in subtree
+        model.links = [link for link in model.links if link.name not in subtree]
+
+        # Remove all joints connected to subtree
+        model.joints = [
+            j
+            for j in model.joints
+            if j.parent not in subtree and j.child not in subtree
+        ]
+
+        # Also remove joint connecting subtree to parent
+        model.joints = [j for j in model.joints if j.child != root_link]
+
+        logger.info(
+            f"Deleted subtree '{root_link}' ({len(subtree)} links) from '{model_id}'"
+        )
+        return True
+
+    def rename_link(
+        self,
+        model_id: str,
+        old_name: str,
+        new_name: str,
+    ) -> bool:
+        """
+        Rename a link.
+
+        Args:
+            model_id: Target model
+            old_name: Current link name
+            new_name: New link name
+
+        Returns:
+            True if renamed
+        """
+        model = self._models.get(model_id)
+        if not model:
+            logger.error(f"Model '{model_id}' not found")
+            return False
+
+        if model.read_only:
+            logger.error(f"Model '{model_id}' is read-only")
+            return False
+
+        # Check for conflicts
+        if model.get_link(new_name):
+            logger.error(f"Link '{new_name}' already exists")
+            return False
+
+        link = model.get_link(old_name)
+        if not link:
+            logger.error(f"Link '{old_name}' not found")
+            return False
+
+        self._save_state()
+
+        # Rename link
+        link.name = new_name
+
+        # Update joint references
+        for joint in model.joints:
+            if joint.parent == old_name:
+                joint.parent = new_name
+            if joint.child == old_name:
+                joint.child = new_name
+
+        # Notify callbacks
+        for callback in self._rename_callbacks:
+            callback(model_id, old_name, new_name)
+
+        logger.info(f"Renamed link '{old_name}' to '{new_name}'")
+        return True
+
+    def rename_joint(
+        self,
+        model_id: str,
+        old_name: str,
+        new_name: str,
+    ) -> bool:
+        """
+        Rename a joint.
+
+        Args:
+            model_id: Target model
+            old_name: Current joint name
+            new_name: New joint name
+
+        Returns:
+            True if renamed
+        """
+        model = self._models.get(model_id)
+        if not model:
+            logger.error(f"Model '{model_id}' not found")
+            return False
+
+        if model.read_only:
+            logger.error(f"Model '{model_id}' is read-only")
+            return False
+
+        # Check for conflicts
+        if model.get_joint(new_name):
+            logger.error(f"Joint '{new_name}' already exists")
+            return False
+
+        joint = model.get_joint(old_name)
+        if not joint:
+            logger.error(f"Joint '{old_name}' not found")
+            return False
+
+        self._save_state()
+        joint.name = new_name
+
+        logger.info(f"Renamed joint '{old_name}' to '{new_name}'")
+        return True
+
+    def modify_joint(
+        self,
+        model_id: str,
+        joint_name: str,
+        **kwargs: Any,
+    ) -> bool:
+        """
+        Modify joint properties.
+
+        Args:
+            model_id: Target model
+            joint_name: Joint to modify
+            **kwargs: Properties to update (origin, axis, limits, dynamics, joint_type)
+
+        Returns:
+            True if modified
+        """
+        model = self._models.get(model_id)
+        if not model:
+            logger.error(f"Model '{model_id}' not found")
+            return False
+
+        if model.read_only:
+            logger.error(f"Model '{model_id}' is read-only")
+            return False
+
+        joint = model.get_joint(joint_name)
+        if not joint:
+            logger.error(f"Joint '{joint_name}' not found")
+            return False
+
+        self._save_state()
+
+        # Update properties
+        if "origin" in kwargs:
+            joint.origin = kwargs["origin"]
+        if "axis" in kwargs:
+            joint.axis = kwargs["axis"]
+        if "limits" in kwargs:
+            joint.limits = kwargs["limits"]
+        if "dynamics" in kwargs:
+            joint.dynamics = kwargs["dynamics"]
+        if "joint_type" in kwargs:
+            joint.joint_type = kwargs["joint_type"]
+
+        logger.info(f"Modified joint '{joint_name}'")
+        return True
+
+    def attach_link(
+        self,
+        model_id: str,
+        parent_link: str,
+        child_link: str,
+        joint_name: str | None = None,
+        joint_type: JointType = JointType.FIXED,
+        origin: Origin | None = None,
+    ) -> bool:
+        """
+        Create a joint attaching two existing links.
+
+        Args:
+            model_id: Target model
+            parent_link: Parent link name
+            child_link: Child link name
+            joint_name: Optional joint name
+            joint_type: Type of joint
+            origin: Joint origin
+
+        Returns:
+            True if attached
+        """
+        model = self._models.get(model_id)
+        if not model:
+            logger.error(f"Model '{model_id}' not found")
+            return False
+
+        if model.read_only:
+            logger.error(f"Model '{model_id}' is read-only")
+            return False
+
+        # Verify links exist
+        if not model.get_link(parent_link):
+            logger.error(f"Parent link '{parent_link}' not found")
+            return False
+        if not model.get_link(child_link):
+            logger.error(f"Child link '{child_link}' not found")
+            return False
+
+        # Check child doesn't already have a parent
+        if model.get_parent(child_link):
+            logger.error(f"Link '{child_link}' already has a parent")
+            return False
+
+        self._save_state()
+
+        # Generate joint name
+        if not joint_name:
+            joint_name = self._generate_unique_name(
+                f"{parent_link}_to_{child_link}_joint",
+                {j.name for j in model.joints},
+            )
+
+        joint = Joint(
+            name=joint_name,
+            joint_type=joint_type,
+            parent=parent_link,
+            child=child_link,
+            origin=origin or Origin(),
+        )
+        model.joints.append(joint)
+
+        logger.info(f"Attached '{child_link}' to '{parent_link}'")
+        return True
+
+    def detach_link(
+        self,
+        model_id: str,
+        link_name: str,
+    ) -> bool:
+        """
+        Detach a link from its parent (remove connecting joint).
+
+        The link becomes a floating root.
+
+        Args:
+            model_id: Target model
+            link_name: Link to detach
+
+        Returns:
+            True if detached
+        """
+        model = self._models.get(model_id)
+        if not model:
+            logger.error(f"Model '{model_id}' not found")
+            return False
+
+        if model.read_only:
+            logger.error(f"Model '{model_id}' is read-only")
+            return False
+
+        joint = self.get_connecting_joint(model_id, link_name)
+        if not joint:
+            logger.warning(f"Link '{link_name}' has no parent joint")
+            return False
+
+        self._save_state()
+
+        model.joints = [j for j in model.joints if j.child != link_name]
+        logger.info(f"Detached link '{link_name}'")
+        return True
+
+    # ============================================================
+    # Batch Operations
+    # ============================================================
+
+    def apply_prefix(
+        self,
+        model_id: str,
+        prefix: str,
+        include_links: bool = True,
+        include_joints: bool = True,
+        include_materials: bool = True,
+    ) -> bool:
+        """
+        Add a prefix to all names in a model.
+
+        Args:
+            model_id: Target model
+            prefix: Prefix to add
+            include_links: Rename links
+            include_joints: Rename joints
+            include_materials: Rename materials
+
+        Returns:
+            True if applied
+        """
+        model = self._models.get(model_id)
+        if not model:
+            logger.error(f"Model '{model_id}' not found")
+            return False
+
+        if model.read_only:
+            logger.error(f"Model '{model_id}' is read-only")
+            return False
+
+        self._save_state()
+
+        # Build name maps
+        link_map: dict[str, str] = {}
+        joint_map: dict[str, str] = {}
+        material_map: dict[str, str] = {}
+
+        if include_links:
+            for link in model.links:
+                link_map[link.name] = prefix + link.name
+
+        if include_joints:
+            for joint in model.joints:
+                joint_map[joint.name] = prefix + joint.name
+
+        if include_materials:
+            for mat_name in model.materials:
+                material_map[mat_name] = prefix + mat_name
+
+        # Apply renames
+        for link in model.links:
+            if link.name in link_map:
+                link.name = link_map[link.name]
+            if link.visual_material and link.visual_material.name in material_map:
+                link.visual_material.name = material_map[link.visual_material.name]
+
+        for joint in model.joints:
+            if joint.name in joint_map:
+                joint.name = joint_map[joint.name]
+            if joint.parent in link_map:
+                joint.parent = link_map[joint.parent]
+            if joint.child in link_map:
+                joint.child = link_map[joint.child]
+
+        # Rename materials in dict
+        new_materials = {}
+        for old_name, mat in model.materials.items():
+            new_name = material_map.get(old_name, old_name)
+            mat.name = new_name
+            new_materials[new_name] = mat
+        model.materials = new_materials
+
+        logger.info(f"Applied prefix '{prefix}' to model '{model_id}'")
+        return True
+
+    def mirror_subtree(
+        self,
+        model_id: str,
+        root_link: str,
+        mirror_axis: str = "y",
+        name_replacements: dict[str, str] | None = None,
+    ) -> list[str]:
+        """
+        Create a mirrored copy of a subtree.
+
+        Useful for creating symmetric limbs (left/right).
+
+        Args:
+            model_id: Target model
+            root_link: Root of subtree to mirror
+            mirror_axis: Axis to mirror across ('x', 'y', or 'z')
+            name_replacements: Name substitutions (e.g., {"left": "right"})
+
+        Returns:
+            List of created link names
+        """
+        # Copy subtree to clipboard
+        if not self.copy_subtree(model_id, root_link):
+            return []
+
+        # Get parent for attachment
+        model = self._models.get(model_id)
+        if not model:
+            return []
+
+        parent = model.get_parent(root_link)
+        if not parent:
+            logger.error("Cannot mirror root link")
+            return []
+
+        # Default replacements for left/right
+        if name_replacements is None:
+            name_replacements = {
+                "left": "right",
+                "right": "left",
+                "Left": "Right",
+                "Right": "Left",
+                "_l_": "_r_",
+                "_r_": "_l_",
+                "_L_": "_R_",
+                "_R_": "_L_",
+            }
+
+        # Paste with mirrored positions
+        self._save_state()
+
+        comp_type, links, joints, materials = self._clipboard[0]
+
+        # Generate mirrored names
+        def mirror_name(name: str) -> str:
+            result = name
+            for old, new in name_replacements.items():
+                result = result.replace(old, new)
+            if result == name:
+                # No replacement found, add suffix
+                result = name + "_mirrored"
+            return result
+
+        # Build name map
+        name_map: dict[str, str] = {}
+        existing_links = {link.name for link in model.links}
+
+        for link in links:
+            new_name = mirror_name(link.name)
+            new_name = self._generate_unique_name(new_name, existing_links)
+            name_map[link.name] = new_name
+            existing_links.add(new_name)
+
+        # Mirror positions
+        axis_idx = {"x": 0, "y": 1, "z": 2}[mirror_axis]
+
+        created_links: list[str] = []
+
+        for link in links:
+            new_link = Link.from_dict(link.to_dict())
+            new_link.name = name_map[link.name]
+
+            # Mirror visual/collision origins
+            if new_link.visual_origin:
+                xyz = list(new_link.visual_origin.xyz)
+                xyz[axis_idx] = -xyz[axis_idx]
+                new_link.visual_origin = Origin(
+                    xyz=tuple(xyz), rpy=new_link.visual_origin.rpy
+                )
+
+            if new_link.collision_origin:
+                xyz = list(new_link.collision_origin.xyz)
+                xyz[axis_idx] = -xyz[axis_idx]
+                new_link.collision_origin = Origin(
+                    xyz=tuple(xyz), rpy=new_link.collision_origin.rpy
+                )
+
+            model.links.append(new_link)
+            created_links.append(new_link.name)
+
+        # Copy and mirror joints
+        for joint in joints:
+            new_joint = Joint.from_dict(joint.to_dict())
+            new_joint.name = mirror_name(joint.name)
+
+            # Update references
+            if joint.parent in name_map:
+                new_joint.parent = name_map[joint.parent]
+            elif joint.child == links[0].name:
+                new_joint.parent = parent  # Attach to same parent
+
+            if joint.child in name_map:
+                new_joint.child = name_map[joint.child]
+
+            # Mirror origin
+            xyz = list(new_joint.origin.xyz)
+            xyz[axis_idx] = -xyz[axis_idx]
+            new_joint.origin = Origin(xyz=tuple(xyz), rpy=new_joint.origin.rpy)
+
+            # Mirror axis for revolute joints
+            if new_joint.joint_type in (JointType.REVOLUTE, JointType.CONTINUOUS):
+                axis = list(new_joint.axis)
+                axis[axis_idx] = -axis[axis_idx]
+                new_joint.axis = tuple(axis)
+
+            model.joints.append(new_joint)
+
+        logger.info(f"Created mirrored subtree with {len(created_links)} links")
+        return created_links
