+++ src/shared/python/model_generation/editor/frankenstein_editor.py
@@ -3,6 +3,11 @@ Frankenstein Editor for component composition.
 
 Allows combining parts from multiple URDF models into a single composite model,
 like building a video game character from different pieces.
+
+This is the main editor class that composes functionality from:
+- editor_clipboard: Clipboard operations (copy link/subtree/material)
+- editor_modifications: Modification operations (delete, rename, attach, mirror)
+- editor_types: Shared data types
 """
 
 from __future__ import annotations
@@ -10,65 +15,25 @@ from __future__ import annotations
 import copy
 import logging
 from collections.abc import Callable
-from dataclasses import dataclass, field
-from enum import Enum
 from pathlib import Path
 from typing import Any
 
 from model_generation.converters.urdf_parser import ParsedModel, URDFParser
 from model_generation.core.types import Joint, JointType, Link, Material, Origin
 
+from .editor_clipboard import ClipboardMixin
+from .editor_modifications import ModificationMixin
+from .editor_types import (
+    ComponentReference,  # noqa: F401
+    ComponentType,
+    EditorState,
+    PendingOperation,  # noqa: F401
+)
+
 logger = logging.getLogger(__name__)
 
 
-class ComponentType(Enum):
-    """Types of components that can be copied."""
-
-    LINK = "link"
-    SUBTREE = "subtree"
-    JOINT = "joint"
-    MATERIAL = "material"
-
-
-@dataclass
-class ComponentReference:
-    """Reference to a component in a model."""
-
-    model_id: str
-    component_type: ComponentType
-    component_name: str
-    # For subtree: the root link name
-    subtree_root: str | None = None
-
-    def __str__(self) -> str:
-        if self.component_type == ComponentType.SUBTREE:
-            return f"{self.model_id}:{self.subtree_root}/*"
-        return f"{self.model_id}:{self.component_name}"
-
-
-@dataclass
-class PendingOperation:
-    """A pending copy/paste operation."""
-
-    operation_type: str  # 'copy_link', 'copy_subtree', 'attach', 'rename', 'delete'
-    source_ref: ComponentReference | None
-    target_model_id: str | None
-    parameters: dict[str, Any] = field(default_factory=dict)
-    preview_links: list[Link] = field(default_factory=list)
-    preview_joints: list[Joint] = field(default_factory=list)
-
-
-@dataclass
-class EditorState:
-    """State of the Frankenstein editor for undo/redo."""
-
-    models: dict[str, ParsedModel]
-    clipboard: list[ComponentReference]
-    operation_history: list[PendingOperation]
-    timestamp: float = 0.0
-
-
-class FrankensteinEditor:
+class FrankensteinEditor(ClipboardMixin, ModificationMixin):
     """
     Editor for composing URDF models from multiple sources.
 
@@ -103,7 +68,7 @@ class FrankensteinEditor:
         urdf_string = editor.export_model("cyborg")
     """
 
-    def __init__(self):
+    def __init__(self) -> None:
         """Initialize the Frankenstein editor."""
         self._models: dict[str, ParsedModel] = {}
         self._parser = URDFParser()
@@ -300,158 +265,6 @@ class FrankensteinEditor:
         return None
 
     # ============================================================
-    # Clipboard Operations
-    # ============================================================
-
-    def copy_link(
-        self,
-        model_id: str,
-        link_name: str,
-        include_joint: bool = True,
-    ) -> bool:
-        """
-        Copy a single link to clipboard.
-
-        Args:
-            model_id: Source model
-            link_name: Link to copy
-            include_joint: Include the connecting joint
-
-        Returns:
-            True if copied
-        """
-        model = self._models.get(model_id)
-        if not model:
-            logger.error(f"Model '{model_id}' not found")
-            return False
-
-        link = model.get_link(link_name)
-        if not link:
-            logger.error(f"Link '{link_name}' not found in '{model_id}'")
-            return False
-
-        links = [Link.from_dict(link.to_dict())]
-        joints = []
-        materials = {}
-
-        if include_joint:
-            joint = self.get_connecting_joint(model_id, link_name)
-            if joint:
-                joints.append(Joint.from_dict(joint.to_dict()))
-
-        # Include materials
-        if link.visual_material:
-            materials[link.visual_material.name] = Material.from_dict(
-                link.visual_material.to_dict()
-            )
-
-        self._clipboard = [(ComponentType.LINK, links, joints, materials)]
-        logger.info(f"Copied link '{link_name}' to clipboard")
-        return True
-
-    def copy_subtree(
-        self,
-        model_id: str,
-        root_link: str,
-    ) -> bool:
-        """
-        Copy a subtree (link and all descendants) to clipboard.
-
-        Args:
-            model_id: Source model
-            root_link: Root link of subtree
-
-        Returns:
-            True if copied
-        """
-        model = self._models.get(model_id)
-        if not model:
-            logger.error(f"Model '{model_id}' not found")
-            return False
-
-        subtree_names = model.get_subtree(root_link)
-        if not subtree_names:
-            logger.error(f"Link '{root_link}' not found in '{model_id}'")
-            return False
-
-        # Copy all links in subtree
-        links = []
-        for name in subtree_names:
-            link = model.get_link(name)
-            if link:
-                links.append(Link.from_dict(link.to_dict()))
-
-        # Copy all joints within subtree
-        joints = []
-        for joint in model.joints:
-            if joint.parent in subtree_names and joint.child in subtree_names:
-                joints.append(Joint.from_dict(joint.to_dict()))
-
-        # Also copy the connecting joint to the subtree root
-        root_joint = self.get_connecting_joint(model_id, root_link)
-        if root_joint:
-            joints.insert(0, Joint.from_dict(root_joint.to_dict()))
-
-        # Collect materials
-        materials = {}
-        for link in links:
-            if link.visual_material:
-                materials[link.visual_material.name] = Material.from_dict(
-                    link.visual_material.to_dict()
-                )
-
-        self._clipboard = [(ComponentType.SUBTREE, links, joints, materials)]
-        logger.info(
-            f"Copied subtree '{root_link}' ({len(links)} links, {len(joints)} joints) to clipboard"
-        )
-        return True
-
-    def copy_material(self, model_id: str, material_name: str) -> bool:
-        """
-        Copy a material definition to clipboard.
-
-        Args:
-            model_id: Source model
-            material_name: Material to copy
-
-        Returns:
-            True if copied
-        """
-        model = self._models.get(model_id)
-        if not model:
-            logger.error(f"Model '{model_id}' not found")
-            return False
-
-        material = model.materials.get(material_name)
-        if not material:
-            logger.error(f"Material '{material_name}' not found in '{model_id}'")
-            return False
-
-        materials = {material_name: Material.from_dict(material.to_dict())}
-        self._clipboard = [(ComponentType.MATERIAL, [], [], materials)]
-        logger.info(f"Copied material '{material_name}' to clipboard")
-        return True
-
-    def get_clipboard_info(self) -> dict[str, Any]:
-        """Get information about clipboard contents."""
-        if not self._clipboard:
-            return {"empty": True}
-
-        comp_type, links, joints, materials = self._clipboard[0]
-        return {
-            "empty": False,
-            "type": comp_type.value,
-            "link_count": len(links),
-            "joint_count": len(joints),
-            "material_count": len(materials),
-            "link_names": [link.name for link in links],
-        }
-
-    def clear_clipboard(self) -> None:
-        """Clear the clipboard."""
-        self._clipboard = []
-
-    # ============================================================
     # Paste Operations
     # ============================================================
 
@@ -494,10 +307,10 @@ class FrankensteinEditor:
         self._save_state()
 
         comp_type, links, joints, materials = self._clipboard[0]
-        created_links = []
+        created_links: list[str] = []
 
         # Build name mapping for renames
-        name_map = {}
+        name_map: dict[str, str] = {}
         existing_links = {link.name for link in model.links}
         existing_joints = {j.name for j in model.joints}
 
@@ -606,581 +419,6 @@ class FrankensteinEditor:
             suffix=suffix,
             joint_type=joint_type,
         )
-
-    # ============================================================
-    # Direct Modifications
-    # ============================================================
-
-    def delete_link(
-        self,
-        model_id: str,
-        link_name: str,
-        reparent_children: bool = True,
-    ) -> bool:
-        """
-        Delete a link from a model.
-
-        Args:
-            model_id: Target model
-            link_name: Link to delete
-            reparent_children: If True, attach children to grandparent
-
-        Returns:
-            True if deleted
-        """
-        model = self._models.get(model_id)
-        if not model:
-            logger.error(f"Model '{model_id}' not found")
-            return False
-
-        if model.read_only:
-            logger.error(f"Model '{model_id}' is read-only")
-            return False
-
-        link = model.get_link(link_name)
-        if not link:
-            logger.error(f"Link '{link_name}' not found")
-            return False
-
-        self._save_state()
-
-        # Get parent
-        parent_name = model.get_parent(link_name)
-        parent_joint = self.get_connecting_joint(model_id, link_name)
-
-        # Get children
-        children = model.get_children(link_name)
-
-        # Reparent children if requested
-        if reparent_children and parent_name:
-            for child in children:
-                child_joint = self.get_connecting_joint(model_id, child)
-                if child_joint:
-                    child_joint.parent = parent_name
-                    # Adjust origin if we have parent joint info
-                    if parent_joint:
-                        # Combine transforms (simplified - just add positions)
-                        px, py, pz = parent_joint.origin.xyz
-                        cx, cy, cz = child_joint.origin.xyz
-                        child_joint.origin = Origin(
-                            xyz=(px + cx, py + cy, pz + cz),
-                            rpy=child_joint.origin.rpy,
-                        )
-        elif not reparent_children:
-            # Delete children recursively
-            for child in children:
-                self.delete_subtree(model_id, child)
-
-        # Remove the link
-        model.links = [link for link in model.links if link.name != link_name]
-
-        # Remove connecting joint
-        model.joints = [j for j in model.joints if j.child != link_name]
-
-        logger.info(f"Deleted link '{link_name}' from '{model_id}'")
-        return True
-
-    def delete_subtree(self, model_id: str, root_link: str) -> bool:
-        """
-        Delete a subtree (link and all descendants).
-
-        Args:
-            model_id: Target model
-            root_link: Root link of subtree to delete
-
-        Returns:
-            True if deleted
-        """
-        model = self._models.get(model_id)
-        if not model:
-            logger.error(f"Model '{model_id}' not found")
-            return False
-
-        if model.read_only:
-            logger.error(f"Model '{model_id}' is read-only")
-            return False
-
-        subtree = model.get_subtree(root_link)
-        if not subtree:
-            logger.error(f"Link '{root_link}' not found")
-            return False
-
-        self._save_state()
-
-        # Remove all links in subtree
-        model.links = [link for link in model.links if link.name not in subtree]
-
-        # Remove all joints connected to subtree
-        model.joints = [
-            j
-            for j in model.joints
-            if j.parent not in subtree and j.child not in subtree
-        ]
-
-        # Also remove joint connecting subtree to parent
-        model.joints = [j for j in model.joints if j.child != root_link]
-
-        logger.info(
-            f"Deleted subtree '{root_link}' ({len(subtree)} links) from '{model_id}'"
-        )
-        return True
-
-    def rename_link(
-        self,
-        model_id: str,
-        old_name: str,
-        new_name: str,
-    ) -> bool:
-        """
-        Rename a link.
-
-        Args:
-            model_id: Target model
-            old_name: Current link name
-            new_name: New link name
-
-        Returns:
-            True if renamed
-        """
-        model = self._models.get(model_id)
-        if not model:
-            logger.error(f"Model '{model_id}' not found")
-            return False
-
-        if model.read_only:
-            logger.error(f"Model '{model_id}' is read-only")
-            return False
-
-        # Check for conflicts
-        if model.get_link(new_name):
-            logger.error(f"Link '{new_name}' already exists")
-            return False
-
-        link = model.get_link(old_name)
-        if not link:
-            logger.error(f"Link '{old_name}' not found")
-            return False
-
-        self._save_state()
-
-        # Rename link
-        link.name = new_name
-
-        # Update joint references
-        for joint in model.joints:
-            if joint.parent == old_name:
-                joint.parent = new_name
-            if joint.child == old_name:
-                joint.child = new_name
-
-        # Notify callbacks
-        for callback in self._rename_callbacks:
-            callback(model_id, old_name, new_name)
-
-        logger.info(f"Renamed link '{old_name}' to '{new_name}'")
-        return True
-
-    def rename_joint(
-        self,
-        model_id: str,
-        old_name: str,
-        new_name: str,
-    ) -> bool:
-        """
-        Rename a joint.
-
-        Args:
-            model_id: Target model
-            old_name: Current joint name
-            new_name: New joint name
-
-        Returns:
-            True if renamed
-        """
-        model = self._models.get(model_id)
-        if not model:
-            logger.error(f"Model '{model_id}' not found")
-            return False
-
-        if model.read_only:
-            logger.error(f"Model '{model_id}' is read-only")
-            return False
-
-        # Check for conflicts
-        if model.get_joint(new_name):
-            logger.error(f"Joint '{new_name}' already exists")
-            return False
-
-        joint = model.get_joint(old_name)
-        if not joint:
-            logger.error(f"Joint '{old_name}' not found")
-            return False
-
-        self._save_state()
-        joint.name = new_name
-
-        logger.info(f"Renamed joint '{old_name}' to '{new_name}'")
-        return True
-
-    def modify_joint(
-        self,
-        model_id: str,
-        joint_name: str,
-        **kwargs: Any,
-    ) -> bool:
-        """
-        Modify joint properties.
-
-        Args:
-            model_id: Target model
-            joint_name: Joint to modify
-            **kwargs: Properties to update (origin, axis, limits, dynamics, joint_type)
-
-        Returns:
-            True if modified
-        """
-        model = self._models.get(model_id)
-        if not model:
-            logger.error(f"Model '{model_id}' not found")
-            return False
-
-        if model.read_only:
-            logger.error(f"Model '{model_id}' is read-only")
-            return False
-
-        joint = model.get_joint(joint_name)
-        if not joint:
-            logger.error(f"Joint '{joint_name}' not found")
-            return False
-
-        self._save_state()
-
-        # Update properties
-        if "origin" in kwargs:
-            joint.origin = kwargs["origin"]
-        if "axis" in kwargs:
-            joint.axis = kwargs["axis"]
-        if "limits" in kwargs:
-            joint.limits = kwargs["limits"]
-        if "dynamics" in kwargs:
-            joint.dynamics = kwargs["dynamics"]
-        if "joint_type" in kwargs:
-            joint.joint_type = kwargs["joint_type"]
-
-        logger.info(f"Modified joint '{joint_name}'")
-        return True
-
-    def attach_link(
-        self,
-        model_id: str,
-        parent_link: str,
-        child_link: str,
-        joint_name: str | None = None,
-        joint_type: JointType = JointType.FIXED,
-        origin: Origin | None = None,
-    ) -> bool:
-        """
-        Create a joint attaching two existing links.
-
-        Args:
-            model_id: Target model
-            parent_link: Parent link name
-            child_link: Child link name
-            joint_name: Optional joint name
-            joint_type: Type of joint
-            origin: Joint origin
-
-        Returns:
-            True if attached
-        """
-        model = self._models.get(model_id)
-        if not model:
-            logger.error(f"Model '{model_id}' not found")
-            return False
-
-        if model.read_only:
-            logger.error(f"Model '{model_id}' is read-only")
-            return False
-
-        # Verify links exist
-        if not model.get_link(parent_link):
-            logger.error(f"Parent link '{parent_link}' not found")
-            return False
-        if not model.get_link(child_link):
-            logger.error(f"Child link '{child_link}' not found")
-            return False
-
-        # Check child doesn't already have a parent
-        if model.get_parent(child_link):
-            logger.error(f"Link '{child_link}' already has a parent")
-            return False
-
-        self._save_state()
-
-        # Generate joint name
-        if not joint_name:
-            joint_name = self._generate_unique_name(
-                f"{parent_link}_to_{child_link}_joint",
-                {j.name for j in model.joints},
-            )
-
-        joint = Joint(
-            name=joint_name,
-            joint_type=joint_type,
-            parent=parent_link,
-            child=child_link,
-            origin=origin or Origin(),
-        )
-        model.joints.append(joint)
-
-        logger.info(f"Attached '{child_link}' to '{parent_link}'")
-        return True
-
-    def detach_link(
-        self,
-        model_id: str,
-        link_name: str,
-    ) -> bool:
-        """
-        Detach a link from its parent (remove connecting joint).
-
-        The link becomes a floating root.
-
-        Args:
-            model_id: Target model
-            link_name: Link to detach
-
-        Returns:
-            True if detached
-        """
-        model = self._models.get(model_id)
-        if not model:
-            logger.error(f"Model '{model_id}' not found")
-            return False
-
-        if model.read_only:
-            logger.error(f"Model '{model_id}' is read-only")
-            return False
-
-        joint = self.get_connecting_joint(model_id, link_name)
-        if not joint:
-            logger.warning(f"Link '{link_name}' has no parent joint")
-            return False
-
-        self._save_state()
-
-        model.joints = [j for j in model.joints if j.child != link_name]
-        logger.info(f"Detached link '{link_name}'")
-        return True
-
-    # ============================================================
-    # Batch Operations
-    # ============================================================
-
-    def apply_prefix(
-        self,
-        model_id: str,
-        prefix: str,
-        include_links: bool = True,
-        include_joints: bool = True,
-        include_materials: bool = True,
-    ) -> bool:
-        """
-        Add a prefix to all names in a model.
-
-        Args:
-            model_id: Target model
-            prefix: Prefix to add
-            include_links: Rename links
-            include_joints: Rename joints
-            include_materials: Rename materials
-
-        Returns:
-            True if applied
-        """
-        model = self._models.get(model_id)
-        if not model:
-            logger.error(f"Model '{model_id}' not found")
-            return False
-
-        if model.read_only:
-            logger.error(f"Model '{model_id}' is read-only")
-            return False
-
-        self._save_state()
-
-        # Build name maps
-        link_map = {}
-        joint_map = {}
-        material_map = {}
-
-        if include_links:
-            for link in model.links:
-                link_map[link.name] = prefix + link.name
-
-        if include_joints:
-            for joint in model.joints:
-                joint_map[joint.name] = prefix + joint.name
-
-        if include_materials:
-            for mat_name in model.materials:
-                material_map[mat_name] = prefix + mat_name
-
-        # Apply renames
-        for link in model.links:
-            if link.name in link_map:
-                link.name = link_map[link.name]
-            if link.visual_material and link.visual_material.name in material_map:
-                link.visual_material.name = material_map[link.visual_material.name]
-
-        for joint in model.joints:
-            if joint.name in joint_map:
-                joint.name = joint_map[joint.name]
-            if joint.parent in link_map:
-                joint.parent = link_map[joint.parent]
-            if joint.child in link_map:
-                joint.child = link_map[joint.child]
-
-        # Rename materials in dict
-        new_materials = {}
-        for old_name, mat in model.materials.items():
-            new_name = material_map.get(old_name, old_name)
-            mat.name = new_name
-            new_materials[new_name] = mat
-        model.materials = new_materials
-
-        logger.info(f"Applied prefix '{prefix}' to model '{model_id}'")
-        return True
-
-    def mirror_subtree(
-        self,
-        model_id: str,
-        root_link: str,
-        mirror_axis: str = "y",
-        name_replacements: dict[str, str] | None = None,
-    ) -> list[str]:
-        """
-        Create a mirrored copy of a subtree.
-
-        Useful for creating symmetric limbs (left/right).
-
-        Args:
-            model_id: Target model
-            root_link: Root of subtree to mirror
-            mirror_axis: Axis to mirror across ('x', 'y', or 'z')
-            name_replacements: Name substitutions (e.g., {"left": "right"})
-
-        Returns:
-            List of created link names
-        """
-        # Copy subtree to clipboard
-        if not self.copy_subtree(model_id, root_link):
-            return []
-
-        # Get parent for attachment
-        model = self._models.get(model_id)
-        if not model:
-            return []
-
-        parent = model.get_parent(root_link)
-        if not parent:
-            logger.error("Cannot mirror root link")
-            return []
-
-        # Default replacements for left/right
-        if name_replacements is None:
-            name_replacements = {
-                "left": "right",
-                "right": "left",
-                "Left": "Right",
-                "Right": "Left",
-                "_l_": "_r_",
-                "_r_": "_l_",
-                "_L_": "_R_",
-                "_R_": "_L_",
-            }
-
-        # Paste with mirrored positions
-        self._save_state()
-
-        comp_type, links, joints, materials = self._clipboard[0]
-
-        # Generate mirrored names
-        def mirror_name(name: str) -> str:
-            result = name
-            for old, new in name_replacements.items():
-                result = result.replace(old, new)
-            if result == name:
-                # No replacement found, add suffix
-                result = name + "_mirrored"
-            return result
-
-        # Build name map
-        name_map = {}
-        existing_links = {link.name for link in model.links}
-
-        for link in links:
-            new_name = mirror_name(link.name)
-            new_name = self._generate_unique_name(new_name, existing_links)
-            name_map[link.name] = new_name
-            existing_links.add(new_name)
-
-        # Mirror positions
-        axis_idx = {"x": 0, "y": 1, "z": 2}[mirror_axis]
-
-        created_links = []
-
-        for link in links:
-            new_link = Link.from_dict(link.to_dict())
-            new_link.name = name_map[link.name]
-
-            # Mirror visual/collision origins
-            if new_link.visual_origin:
-                xyz = list(new_link.visual_origin.xyz)
-                xyz[axis_idx] = -xyz[axis_idx]
-                new_link.visual_origin = Origin(
-                    xyz=tuple(xyz), rpy=new_link.visual_origin.rpy
-                )
-
-            if new_link.collision_origin:
-                xyz = list(new_link.collision_origin.xyz)
-                xyz[axis_idx] = -xyz[axis_idx]
-                new_link.collision_origin = Origin(
-                    xyz=tuple(xyz), rpy=new_link.collision_origin.rpy
-                )
-
-            model.links.append(new_link)
-            created_links.append(new_link.name)
-
-        # Copy and mirror joints
-        for joint in joints:
-            new_joint = Joint.from_dict(joint.to_dict())
-            new_joint.name = mirror_name(joint.name)
-
-            # Update references
-            if joint.parent in name_map:
-                new_joint.parent = name_map[joint.parent]
-            elif joint.child == links[0].name:
-                new_joint.parent = parent  # Attach to same parent
-
-            if joint.child in name_map:
-                new_joint.child = name_map[joint.child]
-
-            # Mirror origin
-            xyz = list(new_joint.origin.xyz)
-            xyz[axis_idx] = -xyz[axis_idx]
-            new_joint.origin = Origin(xyz=tuple(xyz), rpy=new_joint.origin.rpy)
-
-            # Mirror axis for revolute joints
-            if new_joint.joint_type in (JointType.REVOLUTE, JointType.CONTINUOUS):
-                axis = list(new_joint.axis)
-                axis[axis_idx] = -axis[axis_idx]
-                new_joint.axis = tuple(axis)
-
-            model.joints.append(new_joint)
-
-        logger.info(f"Created mirrored subtree with {len(created_links)} links")
-        return created_links
 
     # ============================================================
     # Undo/Redo
@@ -1382,7 +620,7 @@ class FrankensteinEditor:
 
         total_mass = sum(link.inertia.mass for link in model.links)
 
-        joint_types = {}
+        joint_types: dict[str, int] = {}
         for j in model.joints:
             jt = j.joint_type.value
             joint_types[jt] = joint_types.get(jt, 0) + 1
